fix(connector): satisfy managed CLI lint
